+++ core/log/Statistics.php
@@ -2,6 +2,7 @@
 
 namespace Dev4Press\Plugin\CoreActivity\Log;
 
+use DateTime;
 use Dev4Press\Plugin\CoreActivity\Basic\DB;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -80,5 +81,68 @@ class Statistics {
 		}
 
 		return $this->db()->count_entries_by_event_ids( array( $event_id ), $ip, $seconds );
+	}
+
+	public function initial_update() {
+		$statistics = DB::instance()->get_events_statistics();
+
+		$this->save_statistics( $statistics );
+	}
+
+	public function daily_update() {
+		$statistics = coreactivity_settings()->get( 'statistics', 'storage' );
+
+		if ( empty( $statistics ) ) {
+			$this->initial_update();
+		} else {
+			$statistics = json_decode( $statistics, true );
+
+			if ( ! is_array( $statistics ) ) {
+				$this->initial_update();
+			} else {
+				$two = DB::instance()->get_events_statistics( 'two' );
+				$one = DB::instance()->get_events_statistics( 'one' );
+
+				$statistics = $this->update_statistics_array( $statistics, $two );
+				$statistics = $this->update_statistics_array( $statistics, $one );
+
+				$this->save_statistics( $statistics );
+			}
+		}
+	}
+
+	private function update_statistics_array( $statistics, $input ) : array {
+		if ( ! empty( $input ) ) {
+			foreach ( $input as $c => $events ) {
+				if ( ! isset( $statistics[ $c ] ) ) {
+					$statistics[ $c ] = array();
+				}
+
+				foreach ( $events as $e => $days ) {
+					if ( ! isset( $statistics[ $c ][ $e ] ) ) {
+						$statistics[ $c ][ $e ] = array();
+					}
+
+					foreach ( $days as $d => $l ) {
+						if ( ! isset( $statistics[ $c ][ $e ][ $d ] ) ) {
+							$statistics[ $c ][ $e ][ $d ] = 0;
+						}
+
+						if ( $l > $statistics[ $c ][ $e ][ $d ] ) {
+							$statistics[ $c ][ $e ][ $d ] = $l;
+						}
+					}
+				}
+			}
+		}
+
+		return $statistics;
+	}
+
+	private function save_statistics( $statistics ) {
+		coreactivity_settings()->set( 'statistics', json_encode( $statistics ), 'storage' );
+		coreactivity_settings()->set( 'statistics_latest', gmdate( 'Y-m-d' ), 'storage' );
+
+		coreactivity_settings()->save( 'storage' );
 	}
 }