+++ forms/content-tools-updater.php
@@ -12,6 +12,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 			<?php
 
 			include( COREACTIVITY_PATH . 'forms/setup-database.php' );
+			include( COREACTIVITY_PATH . 'forms/setup-statistics.php' );
 
 			?>
         </div>